workload/cli/fixturescmd: relocate from pkg/ccl/workloadccl/cliccl
No Sub-package keeps cloud-storage providers (pulled in via
pkg/workload/fixture) out of generic workloadcli importers. Misleading
"enterprise license required/not required" wording on load/import
short descriptions becomes "via RESTORE"/"via IMPORT".

Epic: none
Release note: None
