+++ pkg/BUILD.bazel
@@ -1064,7 +1064,6 @@ GO_TARGETS = [
     "//pkg/ccl/testutilsccl:testutilsccl",
     "//pkg/ccl/workloadccl/allccl:allccl",
     "//pkg/ccl/workloadccl/allccl:allccl_test",
-    "//pkg/ccl/workloadccl/cliccl:cliccl",
     "//pkg/ccl:ccl",
     "//pkg/changefeed/changefeedpb:changefeedpb",
     "//pkg/changefeed/changefeedpb:checkpointpb",
@@ -3130,6 +3129,7 @@ GO_TARGETS = [
     "//pkg/workload/bank:bank_test",
     "//pkg/workload/bulkingest:bulkingest",
     "//pkg/workload/changefeeds:changefeeds",
+    "//pkg/workload/cli/fixturescmd:fixturescmd",
     "//pkg/workload/cli:cli",
     "//pkg/workload/cli:cli_test",
     "//pkg/workload/conflict:conflict",
