+++ pkg/ccl/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "//pkg/ccl/changefeedccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multitenantccl",
-        "//pkg/ccl/workloadccl/cliccl",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
