+++ pkg/cli/BUILD.bazel
@@ -266,6 +266,7 @@ go_library(
         "//pkg/workload/bank",
         "//pkg/workload/bulkingest",
         "//pkg/workload/cli",
+        "//pkg/workload/cli/fixturescmd",
         "//pkg/workload/debug",
         "//pkg/workload/examples",
         "//pkg/workload/insights",
