+++ pkg/cmd/workload/BUILD.bazel
@@ -7,8 +7,8 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//pkg/ccl/workloadccl/allccl",
-        "//pkg/ccl/workloadccl/cliccl",
         "//pkg/workload/cli",
+        "//pkg/workload/cli/fixturescmd",
     ],
 )
